Allow to get the body as String
In case the body is needed as String (as consumed by the validator),
this is now possible by specifying
* `String` as body type and
* an identity function for transforming the read body value

This solution gives some flexibility on how to transform the String
to the desired value passed to the handler - there may be other cases
than just getting the String.

An alternative solution to support this case would be to handle the
get-string-body case very specifically, but then for other cases
other solutions might have to be introduced. Therefore the more
flexible solution is preferred.

Implementation note: The `Request.withBody` function does not use
kotlin default arguments for the `readValue` parameter (but another
overloading method is added) to keep the existing method signature also
for java users.