Add request to ErrorHandler
So that the error handler can use information of the request for
handling the error, e.g. to log the request uri or some headers.

Note that this is a binary incompatible change (consider to bump the
major version)! Reasoning: Trying to achieve this change in a binary
compatible manner is considered to be adding too much (unnecessary)
code, compared to how easy it is for library users to handle this change.
Additionally, only users with a custom error handler would be affected.